Use random seed for database key on recording logouts
The entry recorded in the domain_mapping_logins table by the logout
handler was just using the unix timestamp as a database key. This caused
duplicate inserts in logging out because a user would be logged out of
multiple sites in the same second. This change adds a random string to
that key to ensure uniqueness (which is the same functionality being
used on logins already.)